# api/openapi.yaml
openapi: 3.0.3
info:
  title: GitStack API
  version: "0.1.0"
  description: >
    API pubblica di GitStack (fonte unica del contratto, decisione D8
    [c_f68d7615da93197a]). Unico punto d'ingresso: il servizio `gateway`.
    UI web, CLI `gs`, skills per agenti e in futuro server MCP usano tutti
    questa stessa API (principio API-first, decisione D7 [c_3780bc749c6ed28d]).

    Il modello dati usa risorse generiche (decisione D15
    [c_4ef209e79c5b2ddb]): oggi la sola risorsa di prova dello scheletro
    end-to-end, domani repository, applicazioni, database e altro senza
    cambiare forma dell'API.

    Ogni versione dell'API vive sotto un prefisso di percorso esplicito
    (`/v1`); `info.version` segue in aggiunta il contratto stesso (SemVer).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /v1
    description: Gateway pubblico di GitStack (dietro Traefik)
...
paths:
...
  /auth/session:
    get:
      operationId: getCurrentSession
      tags: [auth]
      summary: Sessione corrente
      description: >
        Chi e' il chiamante e con quale credenziale (sessione o token
        personale, con i suoi scope). Non richiede scope: basta essere
        autenticati.
      responses:
        "200":
          description: Chiamante autenticato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Elenca gli utenti
      description: Richiede lo scope `read:user`. Il profilo pubblico non include l'email.
      x-required-scopes: ["read:user"]
      parameters:
        - name: q
          in: query
          required: false
          description: Prefisso di username o nome visualizzato.
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: Pagina di utenti.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/UnexpectedError"
    post:
      operationId: createUser
      tags: [users]
      summary: Crea un utente
      description: >
        Solo amministratori, scope `write:user`. Serve a creare utenti umani
        e agenti; non c'e' auto-registrazione.
      x-required-scopes: ["write:user"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: Utente creato.
          headers:
            Location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /users/{username}:
    parameters:
      - $ref: "#/components/parameters/UsernameParam"
    get:
      operationId: getUser
      tags: [users]
      summary: Legge il profilo di un utente
      description: >
        Scope `read:user`. L'email e i campi amministrativi compaiono solo
        per l'utente stesso e per gli amministratori.
      x-required-scopes: ["read:user"]
      responses:
        "200":
          description: Utente trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/UnexpectedError"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Aggiorna il profilo
      description: >
        Scope `write:user`. Solo l'utente stesso (profilo) o un
        amministratore (anche `isAdmin`, `isActive`).
      x-required-scopes: ["write:user"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Utente aggiornato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Elimina un utente
      description: >
        Solo amministratori, scope `write:user`. Elimina a cascata sessioni,
        token, chiavi SSH e membership. Non si puo' eliminare l'ultimo
        amministratore (409 `last_admin`).
      x-required-scopes: ["write:user"]
      responses:
        "204":
          description: Utente eliminato.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /users/{username}/password:
    parameters:
      - $ref: "#/components/parameters/UsernameParam"
    put:
      operationId: changePassword
      tags: [users]
      summary: Cambia la password
      description: >
        Scope `write:user`. L'utente stesso deve fornire `currentPassword`;
        un amministratore no. Con `currentPassword` sbagliata risponde 403
        `forbidden` (il chiamante e' gia' autenticato: un 401 farebbe uscire
        il client). Revoca le altre sessioni dell'utente.
      x-required-scopes: ["write:user"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password cambiata.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: >
        Token personale `gst_...` verificato dal `gateway` tramite identity.
        Stesso schema per CLI `gs`, skills e agenti. Ogni operazione di
        identity indica gli scope richiesti in `x-required-scopes`; le
        sessioni web (cookie) non hanno scope.
    sessionCookie:
      type: apiKey
      in: cookie
      name: gst_session
      description: >
        Sessione web creata da `POST /auth/login` o dal login OIDC. Valore
        opaco; identity ne conserva solo l'hash.
    serviceAuth:
      type: http
      scheme: bearer
      description: >
        Solo interfaccia interna: segreto condiviso di servizio fra gateway,
        core, git e identity. Non e' un token utente.
  parameters:
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UsernameParam:
      name: username
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Name"
...
  schemas:
...
    Error:
      type: object
      description: >
        Formato unico degli errori per tutta l'API pubblica. Codici comuni per
        stato: 400 `bad_request`; 401 `unauthenticated`, `invalid_credentials`;
        403 `forbidden`, `insufficient_scope`, `password_change_required`; 429 `too_many_attempts`; 404 `not_found`; 409 `conflict`
        e varianti specifiche (`already_exists`, `last_admin`, `last_owner`,
        `ssh_key_in_use`, `oidc_identity_unlinked`); 422 `validation_failed`
        (`details.fields`). Le risposte 401 non distinguono mai "utente
        inesistente" da "password errata".
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Codice stabile leggibile da macchina (es. `not_found`).
              example: not_found
            message:
              type: string
              description: Messaggio leggibile da persone, non contrattuale.
            details:
              type: object
              additionalProperties: true
              description: Dettagli opzionali specifici dell'errore.
    Name:
      type: string
      description: >
        Nome breve in minuscolo (username, organizzazione, team, provider):
        lettere minuscole, cifre e trattini, 1-39 caratteri, inizia e finisce
        con un carattere alfanumerico.
      pattern: "^[a-z0-9]([a-z0-9-]{0,37}[a-z0-9])?$"
      minLength: 1
      maxLength: 39
      example: alice
    TokenScope:
      type: string
      description: >
        Catalogo degli scope dei token personali (unico, condiviso da tutti i
        servizi). Le sessioni web non hanno scope: valgono i permessi
        dell'utente. `write:*` include `read:*` dello stesso ambito;
        `admin:org` include `write:org`.
      enum:
        - read:user
        - write:user
        - read:org
        - write:org
        - admin:org
        - read:resource
        - write:resource
...
    User:
      type: object
      description: >
        Profilo utente. `email`, `isAdmin`, `isActive`, `createdAt` sono
        presenti solo per l'utente stesso e per gli amministratori.
      additionalProperties: false
      required: [id, username, kind, displayName]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        username:
          $ref: "#/components/schemas/Name"
        kind:
          type: string
          enum: [human, agent]
        displayName:
          type: string
          maxLength: 128
        bio:
          type: string
          maxLength: 1024
        avatarUrl:
          type: string
          format: uri
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
    CreateUserInput:
      type: object
      additionalProperties: false
      required: [username]
      properties:
        username:
          $ref: "#/components/schemas/Name"
        kind:
          type: string
          enum: [human, agent]
          default: human
        email:
          type: string
          format: email
        displayName:
          type: string
          maxLength: 128
        password:
          type: string
          minLength: 12
          maxLength: 1024
          format: password
          writeOnly: true
          description: Assente per gli agenti (accedono solo con token) e per chi entra solo via OIDC.
        isAdmin:
          type: boolean
          default: false
    UpdateUserInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName:
          type: string
          maxLength: 128
        bio:
          type: string
          maxLength: 1024
        avatarUrl:
          type: string
          format: uri
          nullable: true
        email:
          type: string
          format: email
        isAdmin:
          type: boolean
          description: Solo amministratori.
        isActive:
          type: boolean
          description: Solo amministratori. Disattivare revoca le sessioni e i token.
    UserList:
      type: object
      additionalProperties: false
      required: [items, page, perPage, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
    ChangePasswordInput:
      type: object
      additionalProperties: false
      required: [newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          writeOnly: true
          maxLength: 1024
        newPassword:
          type: string
          minLength: 12
          maxLength: 1024
          format: password
          writeOnly: true
    CurrentSession:
      type: object
      additionalProperties: false
      required: [user, authMethod, mustChangePassword]
      properties:
        user:
          $ref: "#/components/schemas/User"
        authMethod:
          type: string
          enum: [password, oidc, token]
          description: Come e' stato autenticato il chiamante.
        mustChangePassword:
          type: boolean
          description: >
            Vero per l'amministratore creato al primo avvio finche' non cambia
            la password iniziale. Finche' e' vero ogni chiamata autenticata
            risponde 403 `password_change_required`, tranne `GET
            /auth/session`, `POST /auth/logout` e `PUT
            /users/{username}/password` (solo sulla propria utenza).
        scopes:
          type: array
          description: Solo per `authMethod` = `token`.
          items:
            $ref: "#/components/schemas/TokenScope"
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
  responses:
...
    BadRequest:
      description: Richiesta non valida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Token mancante o non valido.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Risorsa non trovata.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflitto con lo stato attuale della risorsa.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnexpectedError:
      description: Errore imprevisto.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: >
        Autenticato ma non autorizzato: scope del token insufficiente
        (`insufficient_scope`, con `details.required`) o permesso mancante
        (`forbidden`), oppure password iniziale ancora da cambiare
        (`password_change_required`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: >
        Richiesta ben formata ma non valida semanticamente (`validation_failed`,
        con `details.fields` = mappa campo -> motivo).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []
  - sessionCookie: []